requirements.txt
# Core dependencies
flet>=0.21.0
aiosqlite>=0.19.0
aiohttp>=3.9.0
packaging>=23.0
psutil>=5.9.0

# Development dependencies
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-asyncio>=0.21.0

# Build dependencies
pyinstaller>=6.0.0

# Optional dependencies for enhanced functionality
requests>=2.31.0
Pillow>=10.0.0

